requirements.txt
asgiref==3.3.1
Django==3.1.3
django-crispy-forms==1.11.0
pkg-resources==0.0.0
pytz==2021.1
sqlparse==0.4.1
